refactor: moved downloadAll for updatePassword from backend to frontend, deleted deleteAll.post.ts
